requirements.txt
python-dotenv
openai
newsapi-python
finnhub-python